A bit of refactoring

+++ src/memory/heap/heap.c
@@ -1,5 +1,5 @@
 #include "heap.h"
-#include "config.h"
+#include "kernel.h"
 #include "status.h"
 #include "memory/memory.h"
 #include <stdbool.h>
@@ -7,9 +7,10 @@
 static int heap_validate_table(void* ptr, void* end, struct heap_table* table)
 {
     int res = 0;
+
     size_t table_size = (size_t)(end - ptr);
     size_t total_blocks = table_size / PEACHOS_HEAP_BLOCK_SIZE;
-    if(table->total != total_blocks)
+    if (table->total != total_blocks)
     {
         res = -EINVARG;
         goto out;
@@ -27,7 +28,8 @@ static bool heap_validate_alignment(void* ptr)
 int heap_create(struct heap* heap, void* ptr, void* end, struct heap_table* table)
 {
     int res = 0;
-    if(!heap_validate_alignment(ptr) || !heap_validate_alignment(end))
+
+    if (!heap_validate_alignment(ptr) || !heap_validate_alignment(end))
     {
         res = -EINVARG;
         goto out;
@@ -38,7 +40,7 @@ int heap_create(struct heap* heap, void* ptr, void* end, struct heap_table* tabl
     heap->table = table;
 
     res = heap_validate_table(ptr, end, table);
-    if(res < 0)
+    if (res < 0)
     {
         goto out;
     }
@@ -52,12 +54,12 @@ out:
 
 static uint32_t heap_align_value_to_upper(uint32_t val)
 {
-    if((val % PEACHOS_HEAP_BLOCK_SIZE) == 0)
+    if ((val % PEACHOS_HEAP_BLOCK_SIZE) == 0)
     {
         return val;
     }
 
-    val = (val - (val % PEACHOS_HEAP_BLOCK_SIZE));
+    val = (val - ( val % PEACHOS_HEAP_BLOCK_SIZE));
     val += PEACHOS_HEAP_BLOCK_SIZE;
     return val;
 }
@@ -70,35 +72,37 @@ static int heap_get_entry_type(HEAP_BLOCK_TABLE_ENTRY entry)
 int heap_get_start_block(struct heap* heap, uint32_t total_blocks)
 {
     struct heap_table* table = heap->table;
-    int bc = 0; // Current block
-    int bs = -1; // Start block
+    int bc = 0;
+    int bs = -1;
 
-    for(size_t i = 0; i < table->total; i++)
+    for (size_t i = 0; i < table->total; i++)
     {
-        if(heap_get_entry_type(table->entries[i]) != HEAP_BLOCK_TABLE_ENTRY_FREE)
+        if (heap_get_entry_type(table->entries[i]) != HEAP_BLOCK_TABLE_ENTRY_FREE)
         {
             bc = 0;
             bs = -1;
             continue;
         }
 
-        if(bs == -1)
+        // If this is the first block
+        if (bs == -1)
         {
             bs = i;
         }
         bc++;
-        if(bc == total_blocks)
+        if (bc == total_blocks)
         {
             break;
         }
     }
 
-    if(bs == -1)
+    if (bs == -1)
     {
         return -ENOMEM;
     }
-
+    
     return bs;
+
 }
 
 void* heap_block_to_address(struct heap* heap, int block)
@@ -106,20 +110,21 @@ void* heap_block_to_address(struct heap* heap, int block)
     return heap->saddr + (block * PEACHOS_HEAP_BLOCK_SIZE);
 }
 
-void heap_mark_blocks_taken(struct heap* heap, int start_block, size_t total_blocks)
+void heap_mark_blocks_taken(struct heap* heap, int start_block, int total_blocks)
 {
-    int end_block = (start_block + total_blocks) - 1;
+    int end_block = (start_block + total_blocks)-1;
+    
     HEAP_BLOCK_TABLE_ENTRY entry = HEAP_BLOCK_TABLE_ENTRY_TAKEN | HEAP_BLOCK_IS_FIRST;
-    if(total_blocks > 1)
+    if (total_blocks > 1)
     {
         entry |= HEAP_BLOCK_HAS_NEXT;
     }
 
-    for(int i = start_block; i <= end_block; i++)
+    for (int i = start_block; i <= end_block; i++)
     {
         heap->table->entries[i] = entry;
         entry = HEAP_BLOCK_TABLE_ENTRY_TAKEN;
-        if(i != end_block - 1)
+        if (i != end_block -1)
         {
             entry |= HEAP_BLOCK_HAS_NEXT;
         }
@@ -129,15 +134,16 @@ void heap_mark_blocks_taken(struct heap* heap, int start_block, size_t total_blo
 void* heap_malloc_blocks(struct heap* heap, uint32_t total_blocks)
 {
     void* address = 0;
+
     int start_block = heap_get_start_block(heap, total_blocks);
-    if(start_block < 0)
+    if (start_block < 0)
     {
         goto out;
     }
 
     address = heap_block_to_address(heap, start_block);
 
-    // Mark block as taken
+    // Mark the blocks as taken
     heap_mark_blocks_taken(heap, start_block, total_blocks);
 
 out:
@@ -147,11 +153,11 @@ out:
 void heap_mark_blocks_free(struct heap* heap, int starting_block)
 {
     struct heap_table* table = heap->table;
-    for(int i = starting_block; i < (int)table->total; i++)
+    for (int i = starting_block; i < (int)table->total; i++)
     {
         HEAP_BLOCK_TABLE_ENTRY entry = table->entries[i];
         table->entries[i] = HEAP_BLOCK_TABLE_ENTRY_FREE;
-        if(!(entry & HEAP_BLOCK_HAS_NEXT))
+        if (!(entry & HEAP_BLOCK_HAS_NEXT))
         {
             break;
         }
@@ -167,7 +173,6 @@ void* heap_malloc(struct heap* heap, size_t size)
 {
     size_t aligned_size = heap_align_value_to_upper(size);
     uint32_t total_blocks = aligned_size / PEACHOS_HEAP_BLOCK_SIZE;
-
     return heap_malloc_blocks(heap, total_blocks);
 }
